added new myTag: NO_WORDPRESS{...} inside this tag, the content does not get filtered by wordpress i.e. shortcodes [someCode params="..."] will literally be displayed as is!
This is needed for the following situation:

we have two columns (very different in height) handled by bootstrap
and we want them to look like this

----------|
|         |
|         |---------|
|         |         |
|         |----------
|         |
-----------

i.e. the second one is to be vertically centered. This is done by a little hack:
<div class="row">
    <div class="col-xs-5 col-md-3 col-lg-1 vcenter">
        <div style="height:10em;border:1px solid #000">Big</div>
    </div><!--
    --><div class="col-xs-5 col-md-7 col-lg-9 vcenter">
        <div style="height:3em;border:1px solid #F00">Small</div>
    </div>
</div>

css:
.vcenter {
    display: inline-block;
    vertical-align: middle;
    float: none;
}

BUT THE COMMENT MUST BE THERE!!!!

wordpress filters this and makes something like
  <!-- <br \>
  </p> !-->
out of it which messes up the rest of the page
--> the new solution looks like

<div class="row">
    <div class="col-xs-5 col-md-3 col-lg-1 vcenter">
        <div style="height:10em;border:1px solid #000">Big</div>
    </div>NO_WORDPRESS{<!--
    -->}<div class="col-xs-5 col-md-7 col-lg-9 vcenter">
        <div style="height:3em;border:1px solid #F00">Small</div>
    </div>
</div>

which prevents wordpress from inserting anything into these hacking areas